requirements.txt
diffusers==0.21.4
transformers
einops
omegaconf
ftfy
tensorboard
modelcards
triton
click
pandas
einops
loralib
regex
tqdm
pytorch_lightning
git+https://github.com/openai/CLIP.git
xformers==0.0.19 